reorganize builder files
