+++ crates/iota-graphql-client/examples/prepare_send_iota_multi.rs
@@ -6,7 +6,7 @@ use std::str::FromStr;
 use anyhow::Result;
 use base64ct::Encoding;
 use iota_graphql_client::Client;
-use iota_transaction_builder::builder::{Res, TransactionBuilder};
+use iota_transaction_builder::{Res, TransactionBuilder};
 use iota_types::{Address, ObjectId};
 
 #[tokio::main]
