+++ crates/iota-sdk-ffi/src/transaction_builder/mod.rs
@@ -3,7 +3,7 @@
 
 use std::sync::{Arc, RwLock, RwLockReadGuard, RwLockWriteGuard};
 
-use iota_transaction_builder::{MovePackageData, builder::PTBArguments};
+use iota_transaction_builder::{MovePackageData, PTBArguments};
 use iota_types::Input;
 
 use crate::{
